+++ docs/changelog.md
@@ -2,6 +2,11 @@
 
 ## Alpha
 
+### 0.0.4
+
+- Improved `repr(layer)`
+- Added `layer.delete`
+
 ### 0.0.3
 
 - Added the `layer.styles` property, a list of `Fill`, `Stroke`, `Shadow` and `InnerShadow` objects, supported by the `RGBA` and `Gradient` structures.